Extract duplicate mockNetworkInterface()
